+++ A4/cd.cpp
@@ -12,9 +12,9 @@ CD::CD(int id_parameter, const char *title_parameter, int artist_id_parameter, c
 
 std::string CD::output() {
 	std::string resultant_string;
-	resultant_string += "\nID: " + std::to_string(id); 
+	resultant_string += "\nID: " + std::to_string(static_cast<long long>(id)); 
 	resultant_string += "\nTitle: " + title;
-	resultant_string += "\nArtist id: " + std::to_string(artist_id);
+	resultant_string += "\nArtist id: " + std::to_string(static_cast<long long>(artist_id));
 	resultant_string += "\nCatalogue: " + catalogue; 
 	length = resultant_string.length();
 	return resultant_string;
